fix(cosmetics): keep the hash helpers out of the search-index graph
Three suites failed to load. cosmetic.service imports ~/server/search-index,
which reaches meilisearch/client — a module that calls pLimit(env...) and
registers prom collectors at import. Importing cosmetic.service from
cosmetic-shop.service and product-badge.service pulled all of that into two
graphs that never had it, past mocks that name modules by hand.

The hash helpers never needed that graph: they want dbWrite, logToAxiom and
getPerceptualHash. Moving them to their own module drops the edge, which
fixes prepaid-membership-jobs with no test change at all.

cosmetic-shop.service still reaches prom/client legitimately, via
orchestrator.service -> workflows -> orchestrator-read-metrics, so those two
mocks now spread the real module with importOriginal instead of listing
exports. A hand-listed mock is coupled to the whole transitive graph of the
thing under test, and nothing warns you when that graph grows.

Verified like-for-like: 3/3 fail before, 11 suites / 53 tests pass after,
matching the same command run against the branch base.
